+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/equipments/Extras.kt
@@ -1,4 +1,4 @@
-package com.hongmyeoun.goldcalc.view.profile.content.equipment
+package com.hongmyeoun.goldcalc.view.profile.content.equipments
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -8,12 +8,9 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Divider
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -21,22 +18,19 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.Bracelet
 import com.hongmyeoun.goldcalc.model.searchedInfo.equipment.CharacterItem
-import com.hongmyeoun.goldcalc.ui.theme.ImageBG
-import com.hongmyeoun.goldcalc.ui.theme.LightGrayBG
 import com.hongmyeoun.goldcalc.ui.theme.LightGrayTransBG
-import com.hongmyeoun.goldcalc.view.profile.normalTextStyle
-import com.hongmyeoun.goldcalc.view.profile.titleTextStyle
 import com.hongmyeoun.goldcalc.view.common.TextChip
 import com.hongmyeoun.goldcalc.view.common.noRippleClickable
+import com.hongmyeoun.goldcalc.view.profile.normalTextStyle
+import com.hongmyeoun.goldcalc.view.profile.titleTextStyle
 import com.hongmyeoun.goldcalc.viewModel.charDetail.EquipmentDetailVM
 
 @Composable
-fun ExtraSimple(
+fun Extra(
     characterEquipment: List<CharacterItem>,
     viewModel: EquipmentDetailVM
 ) {
@@ -48,7 +42,7 @@ fun ExtraSimple(
         horizontalArrangement = Arrangement.SpaceBetween,
         verticalAlignment = Alignment.CenterVertically
     ) {
-        ExtraEquipmentsBracelet(
+        ExtraBracelet(
             modifier = Modifier.weight(1.5f),
             characterEquipment = characterEquipment,
             viewModel = viewModel,
@@ -61,7 +55,7 @@ fun ExtraSimple(
         )
         Spacer(modifier = Modifier.width(8.dp))
 
-        ExtraEquipmentsElixir(
+        ExtraElixir(
             modifier = Modifier.weight(1f),
             characterEquipment = characterEquipment,
             viewModel = viewModel,
@@ -85,7 +79,7 @@ fun ExtraSimple(
 
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
-fun ExtraEquipmentsBracelet(
+fun ExtraBracelet(
     modifier: Modifier,
     characterEquipment: List<CharacterItem>,
     viewModel: EquipmentDetailVM
@@ -152,7 +146,7 @@ fun ExtraEquipmentsBracelet(
 }
 
 @Composable
-fun ExtraEquipmentsElixir(
+fun ExtraElixir(
     modifier: Modifier,
     characterEquipment: List<CharacterItem>,
     viewModel: EquipmentDetailVM
@@ -215,132 +209,4 @@ fun ExtraTrans(
             }
         }
     }
-}
-
-@Composable
-fun BraceletDetailDialog(
-    viewModel: EquipmentDetailVM,
-    characterEquipment: List<CharacterItem>
-) {
-    Dialog(
-        onDismissRequest = { viewModel.onBraDismissRequest() }
-    ) {
-        Column(
-            modifier = Modifier
-                .verticalScroll(rememberScrollState())
-                .background(ImageBG, RoundedCornerShape(16.dp))
-                .padding(16.dp),
-            horizontalAlignment = Alignment.Start
-        ) {
-            Text(
-                text = "팔찌",
-                style = titleTextStyle()
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-            Divider()
-            Spacer(modifier = Modifier.height(8.dp))
-
-            characterEquipment.forEach {
-                when (it) {
-                    is Bracelet -> {
-                        BraDetailTop(viewModel, it)
-
-                        it.extraStats.forEach { (statName, statValue) ->
-                            BraDetailStats(statName, statValue)
-                        }
-
-                        it.specialEffect.forEach { (effectName, tooltip) ->
-                            BraDetailEffects(effectName, tooltip)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-@OptIn(ExperimentalGlideComposeApi::class)
-private fun BraDetailTop(
-    viewModel: EquipmentDetailVM,
-    it: Bracelet
-) {
-    Row(verticalAlignment = Alignment.CenterVertically) {
-        Box(
-            modifier = Modifier
-                .size(56.dp)
-                .background(
-                    brush = viewModel.getItemBG(it.grade),
-                    shape = RoundedCornerShape(8.dp)
-                )
-                .clip(RoundedCornerShape(topStart = 8.dp, topEnd = 8.dp))
-        ) {
-            GlideImage(
-                modifier = Modifier.size(56.dp),
-                model = it.itemIcon,
-                contentDescription = "팔찌 이미지"
-            )
-        }
-        Spacer(modifier = Modifier.width(16.dp))
-
-        Text(
-            text = it.name,
-            style = titleTextStyle()
-        )
-    }
-    Spacer(modifier = Modifier.height(16.dp))
-}
-
-@Composable
-private fun BraDetailStats(statName: String, statValue: String) {
-    Row {
-        TextChip(
-            text = "특성",
-            borderless = true,
-            customBGColor = LightGrayBG
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-
-        TextChip(
-            text = "$statName $statValue",
-            borderless = true,
-            customBGColor = LightGrayBG
-        )
-    }
-    Spacer(modifier = Modifier.height(8.dp))
-}
-
-@Composable
-private fun BraDetailEffects(effectName: String, tooltip: String) {
-    Column {
-        Row {
-            TextChip(
-                text = "효과",
-                borderless = true,
-                customBGColor = LightGrayBG
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-
-            TextChip(
-                text = effectName,
-                borderless = true,
-                customBGColor = LightGrayBG
-            )
-        }
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Box(
-            modifier = Modifier
-                .background(LightGrayBG, RoundedCornerShape(4.dp))
-                .fillMaxWidth()
-                .padding(16.dp)
-        ) {
-            Text(
-                text = tooltip,
-                style = normalTextStyle()
-            )
-        }
-    }
-    Spacer(modifier = Modifier.height(8.dp))
 }
