+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/equipments/SharedUI.kt
@@ -1,0 +1,60 @@
+package com.hongmyeoun.goldcalc.view.profile.content.equipments
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.hongmyeoun.goldcalc.view.common.TextChip
+import com.hongmyeoun.goldcalc.viewModel.charDetail.EquipmentDetailVM
+
+@Composable
+fun UpgradeQualityRow(
+    viewModel: EquipmentDetailVM,
+    grade: String,
+    upgrade: String = "",
+    itemLevel: String = "",
+) {
+    Row(
+        modifier = Modifier.height(24.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        if (upgrade.isNotEmpty()) {
+            Text(
+                text = upgrade,
+                style = TextStyle(
+                    fontSize = 14.sp,
+                    fontWeight = FontWeight.Bold,
+                    textAlign = TextAlign.Start,
+                    color = viewModel.getGradeBG(grade)
+                ),
+            )
+            Spacer(modifier = Modifier.width(4.dp))
+        }
+        if (itemLevel.isNotEmpty()) {
+            if (itemLevel.contains("</FONT>")) {
+                TextChip(itemLevel.substringBefore("</FONT>"))
+            } else {
+                TextChip(itemLevel)
+            }
+            Spacer(modifier = Modifier.width(4.dp))
+        }
+        if (upgrade.isEmpty()) {
+            TextChip(
+                text = grade,
+                borderless = true,
+                customPadding = Modifier.padding(start = 4.dp, end = 4.dp, top = 2.dp, bottom = 2.dp),
+                customBGColor = viewModel.getGradeBG(grade, false)
+            )
+        }
+    }
+}
